Switch to a single header
