+++ src/components/DayPicker.tsx
@@ -3,89 +3,31 @@ import {
   ThemedText,
   ThemedView,
   useThemeColor,
-} from "../Themed";
+} from "./Themed";
 import { theme } from "@/theme";
 import { StyleSheet } from "react-native";
 import { LinearGradient } from "expo-linear-gradient";
 import { ConferenceDay } from "@/consts";
-import {
-  useAnimatedStyle,
-  interpolate,
-  Extrapolation,
-  SharedValue,
-} from "react-native-reanimated";
 
 const backgroundHighlight = { dark: "#1A1A1A", light: "#E6E6E6" };
 
 interface DayPickerProps {
   isDayOne: boolean;
   onSelectDay: (day: ConferenceDay) => void;
-  scrollOffset: SharedValue<number>;
 }
 
-export function DayPicker({
-  isDayOne,
-  onSelectDay,
-  scrollOffset,
-}: DayPickerProps) {
+export function DayPicker({ isDayOne, onSelectDay }: DayPickerProps) {
   const backgroundColor = useThemeColor(theme.color.background);
   const transparentColor = useThemeColor(theme.color.transparent);
 
-  const animatedDayPickerStyle = useAnimatedStyle(() => {
-    const padding = interpolate(
-      scrollOffset.value,
-      [580, 600],
-      [8, 4],
-      Extrapolation.CLAMP,
-    );
-
-    const borderRadius = interpolate(
-      scrollOffset.value,
-      [580, 600],
-      [24, 80],
-      Extrapolation.CLAMP,
-    );
-
-    return {
-      padding,
-      borderRadius,
-    };
-  });
-
-  const animatedDayPickerItemStyle = useAnimatedStyle(() => {
-    const height = interpolate(
-      scrollOffset.value,
-      [580, 600],
-      [48, 40],
-      Extrapolation.CLAMP,
-    );
-
-    const borderRadius = interpolate(
-      scrollOffset.value,
-      [580, 600],
-      [16, 45],
-      Extrapolation.CLAMP,
-    );
-
-    return {
-      height,
-      borderRadius,
-    };
-  });
-
   return (
     <>
-      <ThemedView color={theme.color.background}>
-        <ThemedView
-          style={[styles.dayPicker, animatedDayPickerStyle]}
-          color={backgroundHighlight}
-          animated
-        >
+      <ThemedView color={theme.color.background} style={{ marginTop: 20 }}>
+        <ThemedView style={styles.dayPicker} color={backgroundHighlight}>
           <ThemedPressable
             onPress={() => onSelectDay(ConferenceDay.One)}
             backgroundColor={(!isDayOne && backgroundHighlight) || undefined}
-            style={[styles.dayPickerItem, animatedDayPickerItemStyle]}
-            animated
+            style={styles.dayPickerItem}
           >
             <ThemedText fontWeight="semiBold" fontSize={theme.fontSize16}>
               Day 1
@@ -95,8 +37,7 @@ export function DayPicker({
           <ThemedPressable
             onPress={() => onSelectDay(ConferenceDay.Two)}
             backgroundColor={(isDayOne && backgroundHighlight) || undefined}
-            style={[styles.dayPickerItem, animatedDayPickerItemStyle]}
-            animated
+            style={styles.dayPickerItem}
           >
             <ThemedText fontWeight="semiBold" fontSize={theme.fontSize16}>
               Day 2
@@ -115,16 +56,20 @@ export function DayPicker({
 
 const styles = StyleSheet.create({
   dayPicker: {
-    marginTop: theme.space24,
+    marginTop: theme.space16,
     marginHorizontal: theme.space24,
     flexDirection: "row",
     justifyContent: "space-between",
     alignItems: "center",
+    borderRadius: theme.borderRadius80,
+    padding: theme.space4,
   },
   dayPickerItem: {
     flex: 1,
     justifyContent: "center",
     alignItems: "center",
+    height: 40,
+    borderRadius: theme.borderRadius45,
   },
   fadeGradient: {
     height: 40,
